requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
cloudinary==1.44.1
Django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
drf-nested-routers==0.94.2
environs==14.2.0
idna==3.10
marshmallow==4.0.0
pillow==11.2.1
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
stripe==12.2.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
